+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Changed QasProfile layout and structure
+- Added modifier class `qas-table--mobile`
 ## 1.x.x - 2021-04-29
 
 ### Added